Enum arduino_mkrzero::adc::ctrlb::PRESCALERR

pub enum PRESCALERR {
    DIV4,
    DIV8,
    DIV16,
    DIV32,
    DIV64,
    DIV128,
    DIV256,
    DIV512,
}

Possible values of the field PRESCALER

Variants
DIV4

Peripheral clock divided by 4

DIV8

Peripheral clock divided by 8

DIV16

Peripheral clock divided by 16

DIV32

Peripheral clock divided by 32

DIV64

Peripheral clock divided by 64

DIV128

Peripheral clock divided by 128

DIV256

Peripheral clock divided by 256

DIV512

Peripheral clock divided by 512

Methods

impl PRESCALERR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_div4(&self) -> bool

Checks if the value of the field is DIV4

pub fn is_div8(&self) -> bool

Checks if the value of the field is DIV8

pub fn is_div16(&self) -> bool

Checks if the value of the field is DIV16

pub fn is_div32(&self) -> bool

Checks if the value of the field is DIV32

pub fn is_div64(&self) -> bool

Checks if the value of the field is DIV64

pub fn is_div128(&self) -> bool

Checks if the value of the field is DIV128

pub fn is_div256(&self) -> bool

Checks if the value of the field is DIV256

pub fn is_div512(&self) -> bool

Checks if the value of the field is DIV512