Module arduino_mkrzero::dmac::chctrlb

Channel Control B

Structs

EVIER
Value of the field

EVOER
Value of the field

R
Value read from the register

W
Value to write to the register

_CMDW
Proxy

_EVACTW
Proxy

_EVIEW
Proxy

_EVOEW
Proxy

_LVLW
Proxy

_TRIGACTW
Proxy

_TRIGSRCW
Proxy

Enums

CMDR
Possible values of the field CMD

CMDW
Values that can be written to the field CMD

EVACTR
Possible values of the field EVACT

EVACTW
Values that can be written to the field EVACT

LVLR
Possible values of the field LVL

LVLW
Values that can be written to the field LVL

TRIGACTR
Possible values of the field TRIGACT

TRIGACTW
Values that can be written to the field TRIGACT

TRIGSRCR
Possible values of the field TRIGSRC

TRIGSRCW
Values that can be written to the field TRIGSRC