Enum arduino_mkrzero::port::pmux1_::PMUXOR

pub enum PMUXOR {
    A,
    B,
    C,
    D,
    E,
    F,
    G,
    H,
    _Reserved(u8),
}

Possible values of the field PMUXO

Variants
A

Peripheral function A selected

B

Peripheral function B selected

C

Peripheral function C selected

D

Peripheral function D selected

E

Peripheral function E selected

F

Peripheral function F selected

G

Peripheral function G selected

H

Peripheral function H selected

_Reserved(u8)

Reserved

Methods

impl PMUXOR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_a(&self) -> bool

Checks if the value of the field is A

pub fn is_b(&self) -> bool

Checks if the value of the field is B

pub fn is_c(&self) -> bool

Checks if the value of the field is C

pub fn is_d(&self) -> bool

Checks if the value of the field is D

pub fn is_e(&self) -> bool

Checks if the value of the field is E

pub fn is_f(&self) -> bool

Checks if the value of the field is F

pub fn is_g(&self) -> bool

Checks if the value of the field is G

pub fn is_h(&self) -> bool

Checks if the value of the field is H